+++ src/main/java/com/project/cheerha/domain/viewcount/repository/JobOpeningViewCountRepository.java
@@ -1,26 +1,23 @@
 package com.project.cheerha.domain.viewcount.repository;
 
 import com.project.cheerha.domain.viewcount.entity.JobOpeningViewCount;
-import jakarta.persistence.LockModeType;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 public interface JobOpeningViewCountRepository extends JpaRepository<JobOpeningViewCount, Long> {
 
-    /**
-     * 비관적 락을 사용하여 ViewCount 엔티티를 조회하는 메서드입니다.
-     * 특정 JobOpening ID에 해당하는 JobOpeningViewCount 엔티티를 조회합니다.
-     * @param id 조회할 JobOpening의 ID
-     * @return 비관적 락이 걸린 JobOpeningViewCount 엔티티(Optional 형태)
-     */
-    @Lock(LockModeType.PESSIMISTIC_WRITE)
-    @Query("SELECT b FROM JobOpeningViewCount b WHERE b.jobOpening.id = :id")
-    Optional<JobOpeningViewCount> findWithLockByJobOpeningId(@Param("id") Long id);
+//    /**
+//     * 비관적 락을 사용하여 ViewCount 엔티티를 조회하는 메서드입니다.
+//     * 특정 JobOpening ID에 해당하는 JobOpeningViewCount 엔티티를 조회합니다.
+//     * @param id 조회할 JobOpening의 ID
+//     * @return 비관적 락이 걸린 JobOpeningViewCount 엔티티(Optional 형태)
+//     */
+//    @Query("SELECT b FROM JobOpeningViewCount b WHERE b.jobOpening.id = :id")
+//    Optional<JobOpeningViewCount> findWithLockByJobOpeningId(@Param("id") Long id);
 
     /**
      * 특정 JobOpening ID에 해당하는 ViewCount 엔티티를 조회하는 메서드입니다.
